Fix usage of dev-requirements
The dev-requirements.txt is currently built containing *only*
dev-requirements.in, so both requirements.txt and dev-requirements.txt
will have to be installed.

Alternatively we could combine the two *.in files in the
dev-requirements.txt. At this point, I'm not sure which option is
better.

Remove the --require-hashes from command in the README, as it is implied
by specifying hashes in the requirements file.

Remove the version numbers from the *.in files, as these are not needed.