Add circleci config (#22)
* Add circleci config to run django tests (with nose) and flake8 style checker
  
flake8 is run first, to show failure quickly
Use nose to get a junit test report from django tests.
Use flake8-formatter-junit-xml to get a report  for flake8.

* Add first django and django-webtest tests and a first fixture `testuser`

* Add module for metadata for the testuser fixture

The create_testuser function can be used to recreate the database for
the fixture (e.g. by calling it on the django shell)
and also serves as documentation for the data.

* Fix or disable flake8 warnings

Fix whitespace, linelength etc.
Warnings for scionlab.settings.production/development disabled because
these modules use `from .common import *` as a crude settings
inheritance which flake8 doesn't like.

* Fix usage of dev-requirements in README

The dev-requirements.txt is currently built containing *only*
dev-requirements.in, so both requirements.txt and dev-requirements.txt
will have to be installed.

Remove the version numbers from the *.in files, as these are not needed.