Update all requirements
Prompted by github security alert for django-2.1.5.
Updated all requirements, as this is easier now than later (ha!).

For huey, this is a major version jump (from 1.10 to 2.0). The main
difference for us is that the sqlite backend has moved from contrib to
the main huey module, and that the peewee dependency was removed.