Admin: remove broken UserAS and fix is_ap column
Fix typo in is_ap.

The UserAS admin was broken in that it was just using the AS admin. For
example, creating UserASes would not have worked. Seems fine for now to
just not have support for this. Would be fairly straight forward to add
it back properly if we'd have a use for it.