Resolve various todos, cleanup (#72)
* Fix todo trigger AP deployment, add tests

* Resolve misc todos

* Admin: remove broken UserAS and fix is_ap column

Fix typo in is_ap.

The UserAS admin was broken in that it was just using the AS admin. For
example, creating UserASes would not have worked. Seems fine for now to
just not have support for this. Would be fairly straight forward to add
it back properly if we'd have a use for it.

* Add deploy action to admin & fix deploy task

With None delay, schedule fails (maybe only since huey-2? I think this
had previously worked).

* Add defines module for constants

* Move core modules out of scionlab.util, cleanup

Moved modules certificates, openvpn_config, generate and config_tar from
scionlab.util to scionlab.
Renamed scionlab.util.generate to scionlab.scion_config.

Removed some unused files.

* Fix whitespace flake warning

* Remove underscores in numeric literal

Only supported in python 3.6 apparently; trying to stay compatible with 3.5 for Ubuntu 16.04.