User AS IP input validation. (#70)
* User AS IP input validation.

If the user AS is attaching with VPN, don't care about the IP address. But if
not using VPN, check that the IP is not local, invalid, etc. and that the
IP version of it is accepted by the attachment point.

* Replace IP address in test with a globally reachable one.

Per RFC 5737 192.0.2.0/24 is not globally reachable. Use an address that is.
https://www.iana.org/assignments/iana-ipv4-special-registry/iana-ipv4-special-registry.xhtml

* Changes from review

* UTs

* do not test ip global or loopback if DEBUG==True

* Error tagging the field instead of from form.clean()