Frontend: styling for forms and scripting (#83)
* Use crispy-forms to render forms

Adds some nicer styling for all forms, in particular for the login dialog.
For the UserAS form(s), uses a bit more of crispy's features for the layout of the fields and to include helpful help-text for the fields.

* Add some javascript to

Enable the create/save button only once input has been made to the form,
pop up the "Do you really want to navigate away" dialog if unsaved changes have been made,
dynamically enable/disable the VPN option if available for the selected attachment point, and
hide the IP fields if the VPN option is enabled

* CSS Cleanup, more compact errors, improve handling of invalid fields

Cleanup CSS and make non-field errors look less ugly.
Improve handling of server side validation errors in scripts for user-as
forms.

* Remove now widget_tweaks dependency

* Improve error checks for public ip field

Show error as field error. Clarify error message.
Move protocol version check before detailled check of address (seems like broader
class or error).
Fix css for invalid form feedback with a hack.

* Adapt web-tests to changed HTML

* Bugfix: Dont check for quota when editing a UserAS

Was previously checking for quota during an update to an existing AS.
When the user had maximum allowed number of ASes, this bug would make it
impossible to apply any changes.