Fix race-condition in deployment task start
Before, the deployment task was started while the transaction was still
ongoing. The task checks at the beginning whether it should actually do
run the deployment based on the version number in the DB.
If the task starts up quickly it would abort immediatly, because the
version bump has not been commited to the DB yet.

Fixed by starting the task in transaction.on_commit.