Fix race-condition in deployment task start (#130)
* Fix race-condition in deployment task start

Before, the deployment task was started while the transaction was still
ongoing. The task checks at the beginning whether it should actually do
run the deployment based on the version number in the DB.
If the task starts up quickly it would abort immediatly, because the
version bump has not been commited to the DB yet.

Fixed by starting the task in transaction.on_commit.

* Fix deploy taskqueue tests to work with transaction

* Fix checks for mocked function calls in user as tests

Cant mock deploy_host_config-function any more as this will only be called in
on_commit, which is not triggered in TestCase.
Simply check for the call to AttachmentPoint.trigger_deployment instead.