Skip flaky, irrelevant scrypt harden_runtime test
The test is flaky; the basic harden_runtime implementation seems to be
roughly correct.
However, the random variation in runtime makes, possibly also due to the
multiple scrypt-invocations with separate allocations in the
harden_runtime case, make it hard to test this consistently.

As we don't currently run on this harden_runtime function in production,
and very likely won't in the future (as scrypt is not the default
hasher), we don't want to invest the time to make the tests work reliably.