Use bind IP for services only if appropriate (#147)
Use bind_ip only if the internal_ip matches public_ip; the reason is
that the bind_ip is only really defined for the public_ip. If the
internal_ip is different from public_ip, e.g. if it's 127.0.0.1, then it
doesn't make sense to use bind_ip.
We'd need an internal_bind_ip to represent this correctly -- luckily we
don't seem to need this.