Order entries in user-as list
We've observed that the order of ASes in the list page sometimes changes,
as it is just the order in which the DB returned the entries.
This fix is only for cosmetics, with only 5 ASes it would not be a real
problem to find an AS even if it the order would change on every
refresh.