Check for NULL when iterating over routers/services
When deleting a border router or a service it turns out we are
trying to iterate over a list containing the element which is being
removed. This causes creating a DB lookup over non-existing element
what in effect crashes the engine.

This change introduces a pre-check catching the case where we would
be trying to query a DB over an already non-existing element and
returns an arbitrarily selected value "0" for ID of the element
affected.

Close: #216