Reenable integration tests for VPN, fixes for scionlab-config (#227)
* Reenable integration tests for VPN

Were originally only running locally because it was not possible to run privileged containers in the setup. As we've already been running the integration tests in a circleci VM this is now possible. 

* Update integration test containers to bionic

Our OpenVPN server configuration is not compatible with the OpenVPN version in xenial. This was done intentionally some time ago, and it's acceptable as the infra runs bionic.

* Increase size of dev CA keys, smaller keys not accepted by OpenVPN anymore

* Some tweaks test data to allow simpler tests

* Fixes to scionlab-config (to actually make test work!)

- Actually start VPN server before SCION
- The ccd (obviously) needs to be installed before starting the server.
- Wait for tun0 also for VPN server: after starting VPN client, there was already a very simple wait loop. Also do this for the VPN server to work around the same issue (border router startup requires