Change VPN client conf fileane scheme
scionlab-config alread changed in previous commit.
The idea is to include scionlab in the filename to make it obvious that
we "own" these files. For backwards compatibility with existing
installations of scionlab-config, we leave the name client.conf if there
is only one VPN connection configured (should be good enough for most cases).