Multiple attachment points (#195)
The goal of this change is to let a users attach their UserASes to more than one AttachmentPoint.

* Changes to the UserAS model

Move values for public/bind IP from the UserAS.host to interface level.

Remove the "backup" IP and port fields on UserAS and rely exclusively on the information stored in the Interface objects. This has some small drawbacks (e.g. cannot "recover" public IP when switching to VPN and then back, need to infer "use VPN" state from used IP address), but has the big advantage that no additional per-link state needs to be stored, which simplifies the model.

* Restructure UserAS form logic

Use a form set for the individual provider links. The UserAS form is now a sort of parent form that only includes fields for label and installation type.
Each individual form of the "attachment conf" formset spits out a configuration holder object when saving. All of these are then passed into UserAS.update_attachments(), where the data model is then updated.

* Adapt the html forms and JS

Forms are still rendered using crispy, but it's seemingly not awesome for the now rather complicated formsets.

* Adapted Activate/Deactivate UserAS for multi-ap

* Installation of multiple VPN clients in scionlab-config, change VPN client.conf filename scheme

Adapt name of client.conf files to include scionlab in the new file name scheme to make it obvious that we "own" these files.
For backwards compatibility with existing installations of scionlab-config, we leave the name client.conf if there is only one VPN connection configured (should be good enough for most cases).

* Add a forward port to the VM per interface in the Vagrantfile

* Regenerated test_config_tar/*, using spaces instead of tabs to indent YAML config

Also, fix up some tests (in test_admin) that used the first and _last_
AS for some tests -- as the last AS is a UserAS that now behaves
slightly differently (no default public IP on host object) the tests
would otherwise not work.

* Make crispy fail loudly in DEBUG mode

* Move some form errors to public_ip field because of bad layout

Crispy draws the errors on top of the individual connection forms of the
formset which looks somewhat broken as each sub-form is a "card".
It even looks completely wrong for the "add new attachment point" form,
where the error is rendered _above_ the button to collapse/uncollapse.


Co-authored-by: FR4NK-W <wirzf@inf.ethz.ch>
Co-authored-by: Matthias Frei <matthias.frei@inf.ethz.ch>