Allow accessing home view and API in mainteance mode (#261)
This helps for the next planned maintenance, where we mainly want to
disallow changes on user ASes.

As we may want to keep maintenance mode on for more than a few minutes,
it makes sense to at least show the home page.

Allowing access to the API to fetch host configuration is ok here as
it's a read-only operation. Setting the deployed config version is
obviously not read only, but seems low risk as it's not directly used.
This may simplify the update procedure, as the maintenance mode can be
kept on until all the infrastructure machines are updated.