Include all TRC versions in element configuration
With the "new" control plane PKI in "Manual Mode", it is required to
include at least the base TRC for each ISD.

Previously, this was not required as foreign TRCs were trusted by
(unchecked!) cross signatures.

Also, it is now required to include the entire chain of TRCs to a base
TRC -- it's not allowed (service will fail to start) if only the latest
TRC is present.

As the simplest / safest implementation, we simply include all TRCs of
all ISDs, for now.

Note: at the same time, remove the ISDx/trcs subfolder, as this was only
added for the benefit of scion-pki; as this tool requires per ISD subfolders,
this would not be useful without significantly increasing the width of
the config directory tree. I had only added this for giggles, but now
the fun is over...