Fix hiding IP fields in user AS form
The JS attempts to hide the div-containers holding the IP input fields,
when the VPN option is enabled.
The div container no longer had the class that the JS code was looking
for.
This was caused by the update to django-3, in particular the update
django-crispy-forms.
Fixed by adding an explicit 'hidable' class on the appropriate div.