Add support for fixed links to non-AP ASes
These links can only be modified by admins. As all other changes through
the admin panel, changes to these links do not trigger an automatic
config deployment for the affected hosts.
All applicable link types are allowed, i.e. Infra AS - User AS and User
AS - User AS for both peering and provider links (in either direction).
Note: core links are not an option, user ASes are not core.

The user can only see of the links.

Existing fixed provider links (regardless of which direction) disallow
the user from changing the ISD for this user AS.