Add support for fixed links to non-AP ASes (#302)
Links between user ASes and to non-attachment point ASes can only be
created/modified by admins. For the user, they are read-only and so we call
them "fixed links".  All applicable link types are allowed, i.e. Infra AS -
User AS and User AS - User AS for both peering and provider links (in either
direction). Note: core links are not an option, user ASes are not core.

Existing fixed provider links (regardless of which direction) disallow the user
from changing the ISD for this user AS.

As all other changes through the admin panel, changes to these links do not
trigger an automatic config deployment for the affected hosts.

The user sees a list of the fixed links on the user AS detail page.