Fix prometheus port clashes BR vs CS
The prometheus port of BR and CS could clash.
The BR's prometheus port is set to internal port + 400. Now if the BR's
internal port is 30054, the prometheus port will be 30454 wich is also
the CS's fixed prometheus port.
Using the BR's control port as basis of the prometheus port calculation
fixes this problem (temporarily, I hope), because the range of the
control port overlaps with the CS's service (30254) port and so avoids it.