Fix prometheus port clashes BR vs SD
Similar to previous fix in #321; as the BR's control port could clash
with the sciond port (UDP vs TCP so no problem), the corresponding
prometheus ports would also clash.
So we just avoid the SD port when assigning BR control ports.

Again, this is a hopefully temporary fix (and hence the clunky test).