caddy: fix redirect to www. subdomain (#327)
Add https:// to redirect. Without scheme, the redirect was interpreted
as URI on the same server, leading to a redirect loop.
Note that there is unfortunately no test for this right now, as it would
require faking host names etc. I'm confident this fix works because it's
already running in "production".

Enable logging for the scionlab.org site.