instance_id property on BorderRouter/Service (#346)
In order to get rid of the `control_port` and `internal_port` fields
for the routers, we will want to determine these ports based on the
routers `instance_id`.

The `instance_id` had previously been treated specially and was only
added when fetching the objects to generate the AS topology; this was
done because determining the id based on the PK order is normally
expensive, but we can cheaply annotate the objects when we load all
BorderRouter/Services of an AS anyway.

This change turns the `instance_id` into a "normal" `cached_property` on
BorderRouter/Service, so it is always available. We still do the
annotation optimisation when fetching the objects while generating the
AS topology, by simply overriding the `instance_id` entry in the
object's `__dict__` -- this is exactly how the `cached_property`
operates internally too, so fair game.