Fix config version bump for link active checkbox
The "active" checkbox for the User AS connection to the attachment point
was not bumping the host configuration version when changed.

Fixed by calling `update_active` instead of directly setting the active
field of the link. We have to ensure that the `AttachmentConfForm` model
form does not overwrite this value before that, as this is the reference
value for detecting the `active` state change.