ci: fix python setup (#370)
Cryptography 3.4.7 wheel cannot be easily installed with an old version of pip
(does not support the required wheel tag and whatever that entails).
Update pip before installation to fix this, bump cache key version to
force running this.
Accidentally broken in #366 -- this pip update was there initially but I
removed it later because I thought it was not needed in the newer python
docker image. Due to the cache, the later CI runs did not encounter the
problem.