Add minimal SIG topology entry support
Previously we've had even more minimal support for SIG topology entries
only in User ASes.
Add support for an explicit SIG Service object, to allow configuring the
same thing in the SCIONLab infrastructure ASes.
As before, this is only concerned with adding a topology entry so that
the routers are able to resolve the SIG service address -- actually
configuring the SIG itself is still out of scope and is done manually on
each host.