User Attachment Points (#306)
Allow SCIONLab users to operate attachment points and by that serve as provider for other user ASes.
To run an attachment point, a static, public IP is required. User attachment points may also offer an OpenVPN server.

User attachment points operate in the same way as the existing APs, which have previously been modified to fetch their configuration in a polling based model, instead of previously pushing via SSH, to make this possible.
The health of attachment points is determined by checking whether they are actively polling for configuration updates. APs that fail to do so will no longer be listed. Note that this does not make any guarantees about the health of the APs SCION connectivity.