Use a single router process in attachment points (#381)
Use a single router process in attachment points, vs previously
distributing the interfaces over many routers with max 12 interfaces
each. This was a workaround for some issue in an old version of the
router. Now, this is no longer necessary and running many instances is
harmful for the system's performance.

Remove the automatic splitting of interfaces over many routers.
Once this change is deployed, the script `merge_ap_routers` should be
run to merge and clean up the left router instances.

One notable caveat: the router, as it is implemented now, uses a
blocking write, which can in this configuration potentially slow down
processing of the downstream traffic to the slowest User AS link speed.
This should be fixed in the router implementation, hopefully soon.