Add minimal support for HSR configuration
For the high-speed router (HSR) instances running in the SCIONLab
infrastructure, we need a special configuration (.toml) and run a
different systemd unit.
Instead of adding the very specific configuration settings for the HSR
to the coordinator DB, we just avoid emitting any configuration file for
now, leaving this to the host provisioning (by ansible). So, if the HSR flag
is set for a router, we only add this to the list of wanted systemd
units, nothing else.