API: change versioning in scionlab-config.json manifest
In the scionlab-config.json manifest file, we include a version number
to allow detecting configuration changes cheaply.
This has been useful but one issue with this has been that when changing
the configuration generation logic, e.g. by adding or removing an
optional field somewhere, the configuration version number for all hosts
had to be bumped manually (updating the DB field) or the refreshed
configuration would not be picked up by the hosts.

The new version scheme now adds a generation version number.
The configuration is considered changed and thus regenerated when this
version number changes or, as previously, when the host version counter
is incremented.
The initial value for this generation version number is arbitrarily set
to 13 instead of perhaps more natural 1 -- this should lower the
psychological barrier to increment this number.

This is an API breaking change, thus the API version is incremented
to v4.