Fixup for old version in scionlab-config (#398)
The version sanity for the config info fails for manifest files with the
old versions, so this fails during the upgrade. I had mistakenly assumed
that this sanity check was only applied to the new manifest file. It
turns out that the ConfigInfo does not need the version anyway, so we
can just drop the sanity check altogether.