remove the metrics_port (prometheus) from colibri service.
The colibri service does not yet export any metrics via prometheus;
when the scrappers read the configuration from the api call to
get the topology, they try to read from the metrics_port and
report an instance down.
Remove the appearance of metrics_port from the api call as well
as from the co.toml configuration file.