Generate TRCs in PEM format (#406)
Change the data format of the TRCs from DER to PEM, both in the
generated configuration and in the internal representation in the DB.
Add a data migration to convert all existing TRCs.

Previously, we've generated the TRCs in DER format as this is currently
the default output format of `scion-pki trcs combine`. The (binary) DER
data was stored base64-encoded in the text database field.

According to the Anapayans working on this, the PEM format is should
generally be preferred, and it should become the default format for TRCs
eventually.

The practical motivation for the change is this; the CS persists any
TRCs fetched "over the wire" not only in its sqlite DB, but it also
writes them back to the configuration directory, in PEM format. As our
configuration has up to now contained TRCs in DER, installing updated
configuration using `scionlab-config` after the CS wrote such a TRC file
would lead to an apparent file conflict. Using PEM in our configuration
avoids this.