dev setup: avoid copying dev VPN CA to run/
Refer to the checked-in dev VPN CA directly in the development
configuration, instead of elaborately copying it.

Historical background was that this file was not checked-in but
generated using the initialize_root_ca command. Then it was later added
to git to simplify the setup and make it easier to generate somewhat
deterministic test certificates.
At that time, we used a nose mechanism, `setup_package` to copy the file
to run/ before running the tests.
This doesn't work with django's default test runner. With hindsight, it
really seems that the whole copying business was just unnecessarily
complicated.